+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to show total time after all exercises
 
+### Changed
+
+- Type answer on the same line with the exercise
+
 ## [1.0.1] - 2023-01-06
 
 ### Fixed